Удаление стейтментов, разделение тестов по классам, переименование компонентов в контроллеры
