+++ org.eclipse.rdf4j.jaxrs/pom.xml
@@ -118,7 +118,6 @@
 		<cxf.version>3.2.5</cxf.version>
 		<aries.whiteboard.version>1.0.5</aries.whiteboard.version>
 
-		<rdf4j.version>3.0-SNAPSHOT</rdf4j.version>
 		<rdf4j.server.version>3.1.0-SNAPSHOT</rdf4j.server.version>
 		
 		<lucene.version>7.5.0</lucene.version>
