+++ org.eclipse.rdf4j.jaxrs/ru.agentlab.rdf4j.jaxrs.tests/src/test/java/ru/agentlab/rdf4j/jaxrs/RepositoryControllerTest.java
@@ -1,34 +1,18 @@
 package ru.agentlab.rdf4j.jaxrs;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.greaterThan;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
-import static org.ops4j.pax.exam.CoreOptions.junitBundles;
-import static org.ops4j.pax.exam.CoreOptions.maven;
-import static org.ops4j.pax.exam.CoreOptions.mavenBundle;
-import static org.ops4j.pax.exam.karaf.options.KarafDistributionOption.configureSecurity;
-import static org.ops4j.pax.exam.karaf.options.KarafDistributionOption.editConfigurationFilePut;
-import static org.ops4j.pax.exam.karaf.options.KarafDistributionOption.features;
-import static org.ops4j.pax.exam.karaf.options.KarafDistributionOption.karafDistributionConfiguration;
-import static org.ops4j.pax.exam.karaf.options.KarafDistributionOption.keepRuntimeFolder;
-import static org.ops4j.pax.exam.karaf.options.KarafDistributionOption.logLevel;
 
-import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
 
-import javax.inject.Inject;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 import org.apache.cxf.jaxrs.client.WebClient;
-import org.apache.karaf.itests.KarafTestSupport;
 import org.eclipse.rdf4j.model.Model;
 import org.eclipse.rdf4j.model.util.ModelBuilder;
 import org.eclipse.rdf4j.model.vocabulary.FOAF;
@@ -39,94 +23,25 @@ import org.eclipse.rdf4j.query.TupleQueryResult;
 import org.eclipse.rdf4j.repository.Repository;
 import org.eclipse.rdf4j.repository.RepositoryConnection;
 import org.eclipse.rdf4j.repository.config.ConfigTemplate;
-import org.eclipse.rdf4j.rio.RDFFormat;
-import org.eclipse.rdf4j.rio.Rio;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.ops4j.pax.exam.Configuration;
-import org.ops4j.pax.exam.Option;
-import org.ops4j.pax.exam.ProbeBuilder;
-import org.ops4j.pax.exam.TestProbeBuilder;
 import org.ops4j.pax.exam.junit.PaxExam;
-import org.ops4j.pax.exam.karaf.options.LogLevelOption;
-import org.ops4j.pax.exam.options.MavenArtifactUrlReference;
 import org.ops4j.pax.exam.spi.reactors.ExamReactorStrategy;
 import org.ops4j.pax.exam.spi.reactors.PerClass;
-import org.osgi.framework.Constants;
 
 import ru.agentlab.rdf4j.repository.RepositoryManagerComponent;
 
 
 @RunWith(PaxExam.class)
 @ExamReactorStrategy(PerClass.class)
-public class RepositoryControllerTest extends KarafTestSupport {
-	String ENDPOINT_ADDRESS;
-
-	 @Inject
-	 protected RepositoryManagerComponent manager;
-
-    @Override
-    @Configuration
-    public Option[] config() {
-        MavenArtifactUrlReference karafUrl = maven().groupId("org.apache.karaf").artifactId("apache-karaf").versionAsInProject().type("tar.gz");
-
-        String httpPort = Integer.toString(getAvailablePort(Integer.parseInt(MIN_HTTP_PORT), Integer.parseInt(MAX_HTTP_PORT)));
-        String rmiRegistryPort = Integer.toString(getAvailablePort(Integer.parseInt(MIN_RMI_REG_PORT), Integer.parseInt(MAX_RMI_REG_PORT)));
-        String rmiServerPort = Integer.toString(getAvailablePort(Integer.parseInt(MIN_RMI_SERVER_PORT), Integer.parseInt(MAX_RMI_SERVER_PORT)));
-        String sshPort = Integer.toString(getAvailablePort(Integer.parseInt(MIN_SSH_PORT), Integer.parseInt(MAX_SSH_PORT)));
-        String localRepository = System.getProperty("org.ops4j.pax.url.mvn.localRepository");
-        if (localRepository == null) {
-            localRepository = "";
-        }
-
-        return new Option[] {
-            // enable for remote debugging
-        	//KarafDistributionOption.debugConfiguration("8889", true),
-            karafDistributionConfiguration().frameworkUrl(karafUrl).name("Apache Karaf").unpackDirectory(new File("target/exam")),
-            // enable JMX RBAC security, thanks to the KarafMBeanServerBuilder
-            configureSecurity().disableKarafMBeanServerBuilder(),
-            // configureConsole().ignoreLocalConsole(),
-            keepRuntimeFolder(),
-            logLevel(LogLevelOption.LogLevel.INFO),
-            mavenBundle().groupId("org.awaitility").artifactId("awaitility").versionAsInProject(),
-            mavenBundle().groupId("org.apache.servicemix.bundles").artifactId("org.apache.servicemix.bundles.hamcrest").versionAsInProject(),
-            mavenBundle().groupId("org.apache.karaf.itests").artifactId("common").versionAsInProject(),            
-            features(maven().groupId("ru.agentlab.rdf4j")
-            	.artifactId("ru.agentlab.rdf4j.features").type("xml")
-            	.version("0.0.1-SNAPSHOT"), "ru.agentlab.rdf4j.jaxrs"),
-            //mavenBundle().groupId("org.mockito").artifactId("mockito-core").version("2.23.4"),
-            junitBundles(),
-            editConfigurationFilePut("etc/org.apache.felix.http.cfg", "org.osgi.service.http.port", httpPort),
-            editConfigurationFilePut("etc/org.apache.karaf.management.cfg", "rmiRegistryPort", rmiRegistryPort),
-            editConfigurationFilePut("etc/org.apache.karaf.management.cfg", "rmiServerPort", rmiServerPort),
-            editConfigurationFilePut("etc/org.apache.karaf.shell.cfg", "sshPort", sshPort),
-            editConfigurationFilePut("etc/org.ops4j.pax.url.mvn.cfg", "org.ops4j.pax.url.mvn.localRepository", localRepository)
-        };
-    }
-
-    @ProbeBuilder
-    public TestProbeBuilder probeConfiguration(TestProbeBuilder probe) {
-        System.out.println("TestProbeBuilder gets called");
-        //probe.setHeader(Constants.DYNAMICIMPORT_PACKAGE, "*,org.apache.felix.service.*;status=provisional");
-        probe.setHeader(Constants.IMPORT_PACKAGE, "org.eclipse.rdf4j.query.algebra.evaluation.impl,org.apache.cxf.jaxrs.client");
-        return probe;
-    }
-
+public class RepositoryControllerTest extends Rdf4jJaxrsTestSupport {
+    String ENDPOINT_ADDRESS;
+	
     @Before
     public void init() throws Exception {
         ENDPOINT_ADDRESS = "http://localhost:" + getHttpPort() + "/rdf4j-server/repositories/";
     }
-
-    @Override
-    public String getHttpPort() throws Exception {
-        org.osgi.service.cm.Configuration configuration = configurationAdmin.getConfiguration("org.apache.felix.http", null);
-        if (configuration != null) {
-            return configuration.getProperties().get("org.osgi.service.http.port").toString();
-        }
-        return "8181";
-    }
-
 
     @Test
     public void createQuryAndDeleteNativeRepository_withRepositoryManagerComponent_ShouldWork() throws Exception {
@@ -265,82 +180,5 @@ public class RepositoryControllerTest extends KarafTestSupport {
 		client.close();
 	}
 
-    @Test
-	public void postStatementsShouldWorkOk() throws IOException {
-		String repId = "id1237";
-        assertNull(manager.getRepositoryInfo(repId));
-        Repository repository = manager.getOrCreateRepository(repId, "native-rdfs", null);
-        assertNotNull(repository);
-        assertNotNull(manager.getRepositoryInfo(repId));
-
-		String address = ENDPOINT_ADDRESS + repId + "/statements";
-		WebClient client = WebClient.create(address);
-
-		String file = "/testcases/default-graph-1.ttl";
-		RDFFormat dataFormat = Rio.getParserFormatForFileName(file).orElse(RDFFormat.RDFXML);
-		client.type(dataFormat.getDefaultMIMEType());
-		client.accept(MediaType.WILDCARD);
-
-		InputStream dataStream = RepositoryControllerTest.class.getResourceAsStream(file);
-		assertNotNull(dataStream);
-		assertThat("dataStream.available", dataStream.available(), greaterThan(0));
-
-		System.out.println("POST on address=" + address);
-		Response response = client.post(dataStream);
-		System.out.println("response.status=" + response.getStatusInfo().getReasonPhrase());
-		System.out.println("response.body=" + response.readEntity(String.class));
-		assertEquals(204, response.getStatus());
-		client.close();
-
-		RepositoryConnection repositoryCon = repository.getConnection();
-		assertThat("repositoryCon.size", repositoryCon.size(), equalTo(4L));
-	}
     
-    @Test
-    public void postStatementsToGraphShouldWorkOk() throws IOException {
-        String repId = "id1238";
-        assertNull(manager.getRepositoryInfo(repId));
-        Repository repository = manager.getOrCreateRepository(repId, "native-rdfs", null);
-        assertNotNull(repository);
-        assertNotNull(manager.getRepositoryInfo(repId));
-
-        String address = ENDPOINT_ADDRESS + repId + "/rdf-graphs/graph1";
-        String file = "/testcases/default-graph-1.ttl";
-        RDFFormat dataFormat = Rio.getParserFormatForFileName(file).orElse(RDFFormat.RDFXML);
-        
-        System.out.println("POST statements to graph on address=" + address);
-        WebClient client = WebClient.create(address);
-        client.type(dataFormat.getDefaultMIMEType());
-        client.accept(MediaType.WILDCARD);
-        InputStream dataStream = RepositoryControllerTest.class.getResourceAsStream(file);
-        assertNotNull(dataStream);
-        assertThat("dataStream.available", dataStream.available(), greaterThan(0));
-        Response response = client.post(dataStream);
-        System.out.println("response.status=" + response.getStatusInfo().getReasonPhrase());
-        System.out.println("response.body=" + response.readEntity(String.class));
-        assertEquals(204, response.getStatus());
-        client.close();
-
-        RepositoryConnection repositoryCon = repository.getConnection();
-        assertThat("repositoryCon.size", repositoryCon.size(), equalTo(4L));
-        
-        System.out.println("GET statements from named graph on address=" + address);
-        WebClient client2 = WebClient.create(address);
-        client2.accept(MediaType.WILDCARD);
-        Response response2 = client2.get();
-        System.out.println("response2.status=" + response2.getStatusInfo().getReasonPhrase());
-        System.out.println("response2.body=" + response2.readEntity(String.class));
-        assertEquals(200, response2.getStatus());
-        client2.close();
-        
-        /*String address = ENDPOINT_ADDRESS + repId + "/rdf-graphs/graph1";
-        System.out.println("GET statements from default graph on address=" + address2);
-        WebClient client3 = WebClient.create(address);
-        client3.accept(MediaType.WILDCARD);
-        Response response3 = client3.get();
-        System.out.println("response3.status=" + response3.getStatusInfo().getReasonPhrase());
-        System.out.println("response3.body=" + response3.readEntity(String.class));
-        assertEquals(200, response3.getStatus());
-        client3.close();*/
-    }
 }
