+++ org.eclipse.rdf4j.jaxrs/ru.agentlab.rdf4j.jaxrs.tests/src/test/java/ru/agentlab/rdf4j/jaxrs/StatementsControllerTest.java
@@ -1,0 +1,68 @@
+package ru.agentlab.rdf4j.jaxrs;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.greaterThan;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+
+import org.apache.cxf.jaxrs.client.WebClient;
+import org.eclipse.rdf4j.repository.Repository;
+import org.eclipse.rdf4j.repository.RepositoryConnection;
+import org.eclipse.rdf4j.rio.RDFFormat;
+import org.eclipse.rdf4j.rio.Rio;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.ops4j.pax.exam.junit.PaxExam;
+import org.ops4j.pax.exam.spi.reactors.ExamReactorStrategy;
+import org.ops4j.pax.exam.spi.reactors.PerClass;
+
+@RunWith(PaxExam.class)
+@ExamReactorStrategy(PerClass.class)
+public class StatementsControllerTest extends Rdf4jJaxrsTestSupport {
+    String ENDPOINT_ADDRESS;
+    String address;
+    String file = "/testcases/default-graph-1.ttl";
+    RDFFormat dataFormat = Rio.getParserFormatForFileName(file).orElse(RDFFormat.RDFXML);
+    
+    String repId = "id1237";
+    Repository repository;
+    RepositoryConnection repositoryCon;
+
+    @Before
+    public void init() throws Exception {
+        ENDPOINT_ADDRESS = "http://localhost:" + getHttpPort() + "/rdf4j-server/repositories/";
+        address = ENDPOINT_ADDRESS + repId + "/statements";
+        repository = manager.getOrCreateRepository(repId, "native-rdfs", null);
+        repositoryCon = repository.getConnection();
+    }
+    
+    @After
+    public void cleanup() {
+        repositoryCon.close();
+    }
+
+    @Test
+    public void postStatementsShouldWorkOk() throws IOException {
+        WebClient client = WebClient.create(address);
+        client.type(dataFormat.getDefaultMIMEType());
+        client.accept(MediaType.WILDCARD);
+        InputStream dataStream = RepositoryControllerTest.class.getResourceAsStream(file);
+        assertNotNull(dataStream);
+        assertThat("dataStream.available", dataStream.available(), greaterThan(0));
+        Response response = client.post(dataStream);
+        assertEquals(204, response.getStatus());
+        assertEquals("", response.readEntity(String.class));
+        client.close();
+        
+        assertThat("repositoryCon.size", repositoryCon.size(), equalTo(4L));
+    }
+}
